requirements.txt
python_dotenv
websockets
requests
